+++ NameConverter/NameConverterInterface.php
@@ -25,7 +25,7 @@ interface NameConverterInterface
      * @param string|null          $format
      * @param array<string, mixed> $context
      */
-    public function normalize(string $propertyName/* , ?string $class = null, ?string $format = null, array $context = [] */): string;
+    public function normalize(string $propertyName, ?string $class = null, ?string $format = null, array $context = []): string;
 
     /**
      * Converts a property name to its denormalized value.
@@ -34,5 +34,5 @@ interface NameConverterInterface
      * @param string|null          $format
      * @param array<string, mixed> $context
      */
-    public function denormalize(string $propertyName/* , ?string $class = null, ?string $format = null, array $context = [] */): string;
+    public function denormalize(string $propertyName, ?string $class = null, ?string $format = null, array $context = []): string;
 }