+++ CHANGELOG.md
@@ -5,6 +5,22 @@ CHANGELOG
 ---
 
  * Remove `AbstractNormalizerContextBuilder::withDefaultContructorArguments()`, use `withDefaultConstructorArguments()` instead
+ * Change signature of `NameConverterInterface::normalize()` and `NameConverterInterface::denormalize()` methods:
+
+   Before:
+
+   ```php
+   public function normalize(string $propertyName): string;
+   public function denormalize(string $propertyName): string;
+   ```
+
+   After:
+
+   ```php
+   public function normalize(string $propertyName, ?string $class = null, ?string $format = null, array $context = []): string;
+   public function denormalize(string $propertyName, ?string $class = null, ?string $format = null, array $context = []): string;
+   ```
+ * Remove `AdvancedNameConverterInterface`, use `NameConverterInterface` instead
 
 7.3
 ---